requirements.txt
Flask
selenium
pillow
pytesseract
pandas
requests
openpyxl
